job #11491 cleaning up build CLI
